Se indenta el código, se hacen funciones genéricas y se añade descripción a unas funciones
