+++ BAZProject/Detail/ViewModel/DetailModelView.swift
@@ -7,17 +7,8 @@
 
 import Foundation
 
-class DetailModelView {
+final class DetailModelView {
     
     var dataArray: Movie?
     
-    /// La funcion getImage retorna la informacion de una imagen
-    /// - Parameter urlImage: String con la url de la imagen a descargar
-    /// - Returns: La infotmacion de la imagen descargada de tipo Data
-    func getImage(urlImage: String) -> Data {
-        let urlIm = "https://image.tmdb.org/t/p/w500\(urlImage)"
-        guard let urlData = URL(string: urlIm),
-              let data = try? Data(contentsOf: urlData) else { return Data() }
-        return data
-    }
 }
