+++ BAZProject/Home/Cells/MovieTableViewCell.swift
@@ -7,7 +7,7 @@
 
 import UIKit
 import SkeletonView
-class MovieTableViewCell: UITableViewCell, GenericCellMovie {
+final class MovieTableViewCell: UITableViewCell, GenericCellMovie {
     
     @IBOutlet weak var infoStackView: UIStackView!
     @IBOutlet weak var imageViewMovie: UIImageView!
@@ -40,7 +40,7 @@ class MovieTableViewCell: UITableViewCell, GenericCellMovie {
     
     func configure() {
         guard let movieModel = movieModel else { return }
-        let imageMovie = UIImage(data: self.movieViewModel.getImage(urlImage: movieModel.posterPath)) ?? UIImage(named: "poster")
+        let imageMovie = UIImage(data: GenericFunctions.getImage(urlImage: movieModel.posterPath)) ?? UIImage(named: "poster")
         self.configureData(titulo: movieModel.title, originalTitle: movieModel.originalTitle, calificacion: movieModel.voteAverage, image: imageMovie ?? UIImage())
     }
