+++ BAZProject/Home/Model/APIError.swift
@@ -13,6 +13,7 @@ enum APIError: Error {
     case invalidData
     case responseUnsuccessful
     case jsonParsingFailure
+    
     var localizedDescription: String {
         switch self {
         case .requestFailed: return "Request Failed"
