+++ BAZProject/Search/View/SearchViewControllerExtension.swift
@@ -1,0 +1,36 @@
+//
+//  SearchViewControllerExtension.swift
+//  BAZProject
+//
+//  Created by Alan Emiliano Ramirez Ayala on 14/11/22.
+//
+
+import UIKit
+
+extension SearchViewController: UICollectionViewDelegate, UICollectionViewDataSource {
+    
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        self.searchViewModel.dataArray.count
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        let cell = self.CollectionSearchView.dequeueReusableCell(withReuseIdentifier: "SearchCollectionViewCell", for: indexPath) as! SearchCollectionViewCell
+        cell.setUpSkeleton()
+        cell.image.showAnimatedGradientSkeleton()
+        cell.title.showAnimatedGradientSkeleton()
+        DispatchQueue.main.asyncAfter(deadline: .now()) { [self] in
+            cell.image.hideSkeleton()
+            cell.title.hideSkeleton()
+            cell.movieModel = searchViewModel.dataArray[indexPath.row]
+        }
+        return cell
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
+        let vc = UIStoryboard(name: "Main", bundle: nil).instantiateViewController(withIdentifier: "DetailViewController") as! DetailViewController
+        vc.detailViewModel.dataArray = searchViewModel.dataArray[indexPath.row]
+        vc.modalPresentationStyle = .fullScreen
+        self.navigationController?.pushViewController(vc, animated: true)
+    }
+    
+}
