+++ BAZProject/Search/ViewModel/SearchViewModel.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-class SearchViewModel: PrincipalViewModel {
+final class SearchViewModel: PrincipalViewModel {
      
     var refreshData = { () -> () in }
     
@@ -24,13 +24,5 @@ class SearchViewModel: PrincipalViewModel {
     func getInfo(_ api: MovieFeed) {
         self.dataArray = self.movieApiDelegate?.getMovies(api) ?? []
     }
-    /// La funcion getImage retorna la informacion de una imagen
-    /// - Parameter urlImage: String con la url de la imagen a descargar
-    /// - Returns: La infotmacion de la imagen descargada de tipo Data
-    func getImage(urlImage: String) -> Data {
-        let urlIm = "https://image.tmdb.org/t/p/w500\(urlImage)"
-        guard let urlData = URL(string: urlIm),
-              let data = try? Data(contentsOf: urlData) else { return Data() }
-        return data
-    }
+    
 }
